requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-admin==2.0.2
django-consul==0.4.0
django-excel-response2==3.0.6
django-rest-swagger==2.2.0
django-six==1.0.5
djangorestframework==3.14.0
drf-yasg==1.21.7
excel-base==1.0.4
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
protobuf==4.21.12
pytesseract==0.3.10
python-consul==1.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
screen==1.0.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.4
xlwt==1.3.0
